Submission post michael
So, I got the sound working well, didn't have time to do the menu and the fade to black covers the UI for restarting the level. The level end needs to be done as well. The art looks nice, ladders might need to be stacked in order for them to work